refactor: organise loops JS logic into folders
