+++ nodes/web/helpers/loops/main.js
@@ -1,0 +1,4 @@
+import { NODE_COLORS } from "./style.js";
+import { excludeLoops, updateInputsOutputs } from "./utils.js";
+
+export { excludeLoops, NODE_COLORS, updateInputsOutputs };
