+++ nodes/web/helpers/loops/style.js
@@ -1,0 +1,3 @@
+const NODE_COLORS = { nodeColor: "#2a363b", nodeBgColor: "#3f5159" };
+
+export { NODE_COLORS };
